function_app/requirements.txt
azure-functions
azure-storage-common
azure-storage-blob
azure-cosmosdb-table
numpy
opencv-python